iris/sumi: widen ring workspace size to long
initBuffers computed 'int size = nelems_ * type_size_' while finalizeBuffers
frees with a long; the int overflows first at >2GB workspace, mismatching the
free. allocateWorkspace/memcopy already take uint64_t, so widen the local.

+++ src/sst/elements/iris/sumi/reduce_scatter.cc
@@ -67,7 +67,7 @@ HalvingReduceScatterActor::initBuffers()
 {
   void* dst = result_buffer_;
   void* src = send_buffer_;
-  int size = nelems_ * type_size_;
+  long size = static_cast<long>(nelems_) * type_size_; // matches finalizeBuffers
   result_buffer_ = dst;
   if (src != dst)
     my_api_->memcopy(dst, src, size);
